Refactor by moving the swagger docs to a different folder and file and importing it.
